// src/app/models/tgtInfo.model.ts
export interface TgtInfo{
    
         Month:  string;
         SerTarget?: number;
         IFCTarget?: number;
         EmailTarget?: number;
         PSFTarget?: number;
         SMRTarget?: number;
         SerAch?: number;
         IFCAch?: number;  
         EmailAch?: number;
         PSFAch?: number;
         SMRAch?: number;
    }